backend/requirements.txt
# Django core
Django==5.1.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.1

# Database
psycopg[binary]==3.2.6

# Image handling
Pillow==11.1.0

# Async tasks
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0

# Production server
gunicorn==23.0.0

# Environment & config
python-decouple==3.8

# Static files
whitenoise==6.9.0

# iyzico payment
iyzipay==1.0.45
requests==2.32.5

# Rate limiting
django-ratelimit==4.1.0

# PDF generation
reportlab==4.2.5
weasyprint==62.3

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-asyncio==0.24.0
factory-boy==3.3.1
faker==33.1.0
